requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.10
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
black==22.8.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.50.0
django-debug-toolbar==3.7.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
isort==5.10.1
jsonschema==4.16.0
kombu==5.2.4
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pylint==2.15.3
pyparsing==3.0.9
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
types-cryptography==3.3.23
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
